fix: Split server-only package code out to avoid bundling `node:async_hooks` on the client
