+++ packages/kit/src/exports/internal/index.js
@@ -62,27 +62,4 @@ export class ActionFailure {
 	}
 }
 
-export {
-	with_request_store,
-	getRequestEvent,
-	get_request_store,
-	try_get_request_store
-} from './event.js';
-
 export { validate_remote_functions } from './remote-functions.js';
-
-/**
- * @template {{ tracing: { enabled: boolean, root: import('@opentelemetry/api').Span, current: import('@opentelemetry/api').Span } }} T
- * @param {T} event_like
- * @param {import('@opentelemetry/api').Span} current
- * @returns {T}
- */
-export function merge_tracing(event_like, current) {
-	return {
-		...event_like,
-		tracing: {
-			...event_like.tracing,
-			current
-		}
-	};
-}
