+++ packages/kit/src/runtime/app/server/remote/form.js
@@ -1,6 +1,6 @@
 /** @import { RemoteForm } from '@sveltejs/kit' */
 /** @import { RemoteInfo, MaybePromise } from 'types' */
-import { get_request_store } from '@sveltejs/kit/internal';
+import { get_request_store } from '@sveltejs/kit/internal/server';
 import { check_experimental, run_remote_function } from './shared.js';
 
 /**
