+++ packages/kit/src/runtime/server/page/actions.js
@@ -3,14 +3,8 @@
 import * as devalue from 'devalue';
 import { DEV } from 'esm-env';
 import { json } from '@sveltejs/kit';
-import {
-	HttpError,
-	Redirect,
-	ActionFailure,
-	SvelteKitError,
-	with_request_store,
-	merge_tracing
-} from '@sveltejs/kit/internal';
+import { HttpError, Redirect, ActionFailure, SvelteKitError } from '@sveltejs/kit/internal';
+import { with_request_store, merge_tracing } from '@sveltejs/kit/internal/server';
 import { get_status, normalize_error } from '../../../utils/error.js';
 import { is_form_content_type, negotiate } from '../../../utils/http.js';
 import { handle_error_and_jsonify } from '../utils.js';
