+++ packages/kit/src/runtime/server/utils.js
@@ -1,6 +1,7 @@
 import { DEV } from 'esm-env';
 import { json, text } from '@sveltejs/kit';
-import { HttpError, with_request_store } from '@sveltejs/kit/internal';
+import { HttpError } from '@sveltejs/kit/internal';
+import { with_request_store } from '@sveltejs/kit/internal/server';
 import { coalesce_to_error, get_message, get_status } from '../../utils/error.js';
 import { negotiate } from '../../utils/http.js';
 import { fix_stack_trace } from '../shared-server.js';
