+++ packages/kit/tsconfig.json
@@ -13,6 +13,7 @@
 			"@sveltejs/kit/node": ["./src/exports/node/index.js"],
 			"@sveltejs/kit/node/polyfills": ["./src/exports/node/polyfills.js"],
 			"@sveltejs/kit/internal": ["./src/exports/internal/index.js"],
+			"@sveltejs/kit/internal/server": ["./src/exports/internal/server.js"],
 			"$app/server": ["./src/runtime/app/server/index.js"],
 			// internal use only
 			"types": ["./src/types/internal.d.ts"]
